+++ src/case/button/item.singleselect.js
@@ -21,8 +21,6 @@ BI.SingleSelectItem = BI.inherit(BI.BasicButton, {
             text: o.text,
             keyword: o.keyword,
             value: o.value,
-            title: o.title || o.text,
-            warningTitle: o.warningTitle,
             py: o.py
         });
     },